python_backend/requirements.txt
flask-socketio==5.3.6
python-socketio==5.11.0
websockets==12.0
python-dotenv==1.0.0
